requirements.txt
streamlit==1.35.0
easyocr==1.7.1
pytesseract==0.3.10
opencv-python==4.9.0.80
Pillow==10.3.0
numpy==1.26.4
